fixing structure and wallet functionality
