+++ Landing/jsconfig.json
@@ -1,0 +1,11 @@
+{
+	"compilerOptions": {
+		"baseUrl": "./",
+		"paths": {
+			"@hooks/*": ["_hooks/*"],
+			"@includes/*": ["_includes/*"],
+			"@layouts/*": ["_layouts/*"],
+			"@app": ["pages/_app.js"],
+		}
+	}
+}
